Add BlocMenu
Currently, it is only possible to update a bloc by typing prefixes. This shortcut can be useful but needs to remains a shortcut. To be able to really transform blocs, we add a menu on the left of a bloc.

We also update some tests that were invalidated due to the change of DOM.