Add pagination system
To avoid loading all images in database, we add a pagination loading.

The solution avoids infinite scroll (dark pattern) and prefer a specific button to load more images.

The solution uses a cursor-like pagination and not a numbered pagination. The search should better help the user instead of a pagination. Also, adding results instead of replacing it is not quite compatible with a numbered pagination.